requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.5
djangorestframework==3.12.2
flake8==3.8.4
importlib-metadata==3.4.0
iniconfig==1.1.1
mccabe==0.6.1
packaging==20.8
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.1
pytz==2020.5
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
zipp==3.4.0
pytest-django==4.1.0 
pytest-cov==2.11.1 
